pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-synthesis"
version = "0.1.0"
description = "Deterministic news pipeline scaffold for the Excalibur news repo."
requires-python = ">=3.11"
dependencies = [
  "feedparser>=6,<7",
  "pydantic>=2,<3",
  "python-dotenv>=1,<2",
  "PyYAML>=6,<7",
  "requests>=2,<3",
  "typer>=0.12,<1",
]

[project.scripts]
news-synthesis = "news_synthesis.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]